chore: reafactor for mvcc compliance
